Backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
python-multipart
spacy
nltk
python-jose[cryptography]
passlib[bcrypt]
python-decouple
statsmodels
